makeDirをDodontoF::Utilsに分離した
